fix(integrations): Export container renderers from a dedicated export path to fix bundling issues
