+++ packages/integrations/mdx/src/container-renderer.ts
@@ -1,0 +1,8 @@
+import type { AstroRenderer } from 'astro';
+
+export function getContainerRenderer(): AstroRenderer {
+	return {
+		name: 'astro:jsx',
+		serverEntrypoint: '@astrojs/mdx/server.js',
+	};
+}
