+++ packages/integrations/solid/src/container-renderer.ts
@@ -1,0 +1,9 @@
+import type { AstroRenderer } from 'astro';
+
+export function getContainerRenderer(): AstroRenderer {
+	return {
+		name: '@astrojs/solid-js',
+		clientEntrypoint: '@astrojs/solid-js/client.js',
+		serverEntrypoint: '@astrojs/solid-js/server.js',
+	};
+}
