+++ packages/integrations/solid/src/index.ts
@@ -1,6 +1,7 @@
 import type { AstroIntegration, AstroIntegrationLogger, AstroRenderer } from 'astro';
 import type { PluginOption, Plugin } from 'vite';
 import solid, { type Options as ViteSolidPluginOptions } from 'vite-plugin-solid';
+import { getContainerRenderer as getContainerRendererImpl } from './container-renderer.js';
 
 // TODO: keep in sync with https://github.com/thetarnav/solid-devtools/blob/main/packages/main/src/vite/index.ts#L7
 type DevtoolsPluginOptions = {
@@ -56,15 +57,15 @@ function getViteConfiguration(
 	return { plugins };
 }
 
-function getRenderer(): AstroRenderer {
-	return {
-		name: '@astrojs/solid-js',
-		clientEntrypoint: '@astrojs/solid-js/client.js',
-		serverEntrypoint: '@astrojs/solid-js/server.js',
-	};
+/**
+ * @deprecated Import `getContainerRenderer` from `@astrojs/solid-js/container-renderer` instead.
+ */
+export function getContainerRenderer(): AstroRenderer {
+	console.warn(
+		'[@astrojs/solid-js] Importing `getContainerRenderer` from `@astrojs/solid-js` is deprecated. Import it from `@astrojs/solid-js/container-renderer` instead.',
+	);
+	return getContainerRendererImpl();
 }
-
-export { getRenderer as getContainerRenderer };
 
 export interface Options extends Pick<ViteSolidPluginOptions, 'include' | 'exclude'> {
 	devtools?: boolean;
@@ -86,7 +87,7 @@ export default function (options: Options = {}): AstroIntegration {
 					!!options.devtools && command === 'dev',
 				);
 
-				addRenderer(getRenderer());
+				addRenderer(getContainerRendererImpl());
 				updateConfig({
 					vite: getViteConfiguration(options, devtoolsPlugin),
 				});
