+++ packages/integrations/svelte/src/container-renderer.ts
@@ -1,0 +1,9 @@
+import type { AstroRenderer } from 'astro';
+
+export function getContainerRenderer(): AstroRenderer {
+	return {
+		name: '@astrojs/svelte',
+		clientEntrypoint: '@astrojs/svelte/client.js',
+		serverEntrypoint: '@astrojs/svelte/server.js',
+	};
+}
