+++ packages/integrations/vue/src/container-renderer.ts
@@ -1,0 +1,9 @@
+import type { AstroRenderer } from 'astro';
+
+export function getContainerRenderer(): AstroRenderer {
+	return {
+		name: '@astrojs/vue',
+		clientEntrypoint: '@astrojs/vue/client.js',
+		serverEntrypoint: '@astrojs/vue/server.js',
+	};
+}
