+++ packages/integrations/vue/src/index.ts
@@ -6,6 +6,7 @@ import { MagicString } from '@vue/compiler-sfc';
 import type { AstroIntegration, AstroRenderer, HookParameters } from 'astro';
 import type { EnvironmentOptions, Plugin, PluginOption } from 'vite';
 import type { VitePluginVueDevToolsOptions } from 'vite-plugin-vue-devtools';
+import { getContainerRenderer as getContainerRendererImpl } from './container-renderer.js';
 
 const VIRTUAL_MODULE_ID = 'virtual:astro:vue-app';
 const RESOLVED_VIRTUAL_MODULE_ID = `\0${VIRTUAL_MODULE_ID}`;
@@ -16,14 +17,6 @@ interface Options extends VueOptions {
 	devtools?: boolean | Omit<VitePluginVueDevToolsOptions, 'appendTo'>;
 }
 
-function getRenderer(): AstroRenderer {
-	return {
-		name: '@astrojs/vue',
-		clientEntrypoint: '@astrojs/vue/client.js',
-		serverEntrypoint: '@astrojs/vue/server.js',
-	};
-}
-
 function getJsxRenderer(): AstroRenderer {
 	return {
 		name: '@astrojs/vue (jsx)',
@@ -32,7 +25,15 @@ function getJsxRenderer(): AstroRenderer {
 	};
 }
 
-export { getRenderer as getContainerRenderer };
+/**
+ * @deprecated Import `getContainerRenderer` from `@astrojs/vue/container-renderer` instead.
+ */
+export function getContainerRenderer(): AstroRenderer {
+	console.warn(
+		'[@astrojs/vue] Importing `getContainerRenderer` from `@astrojs/vue` is deprecated. Import it from `@astrojs/vue/container-renderer` instead.',
+	);
+	return getContainerRendererImpl();
+}
 
 function virtualAppEntrypoint(options?: Options): Plugin {
 	let isBuild: boolean;
@@ -191,7 +192,7 @@ export default function (options?: Options): AstroIntegration {
 		name: '@astrojs/vue',
 		hooks: {
 			'astro:config:setup': async ({ addRenderer, updateConfig, command }) => {
-				addRenderer(getRenderer());
+				addRenderer(getContainerRendererImpl());
 				if (options?.jsx) {
 					addRenderer(getJsxRenderer());
 				}
